Debugging "split_data"

+++ translation.py
@@ -23,7 +23,7 @@ def creat_pairs(lines):
         english, spanish, other = line.split("\t") # creating english and spanish words by camma
         spanish = "[start]" + spanish + "[end]" # Add [start] and [end] to spanish words
         text_pairs.append((english, spanish))
-        return text_pairs
+    return text_pairs
 text_pairs = creat_pairs(lines)
 
 # Split data to train-test-validation
@@ -32,9 +32,9 @@ test_per = 0.15
 val_per = 0.15
 def split_data(text_pairs):
     random.shuffle(text_pairs) # shuffle text data
-    num_train_data = train_per * (len(text_pairs)) # number of train data
+    num_train_data = int(train_per * (len(text_pairs))) # number of train data
     #num_test_data = test_per * (len(text_pairs)) # number of test data
-    num_val_data = val_per * (len(text_pairs)) # number of validation data
+    num_val_data = int(val_per * (len(text_pairs))) # number of validation data
     train_pairs = text_pairs[:num_train_data] # train data from 0 to the end of train data numbers
     val_pairs = text_pairs[num_train_data:num_train_data + num_val_data] # validation data from train numbers to the sum of train and validation numbers
     test_pairs = text_pairs[num_train_data + num_val_data:] # test data from sum of train and validation numbers to the end
